Break out command line functions to utilities file Fixes #19
