+++ .DBConfig
@@ -1,0 +1,7 @@
+{
+    "host": "localhost",
+    "port": "5432",
+    "user": "tuskorm",
+    "password": "tuskorm",
+    "database": "tuskorm_test"
+}
