+++ .gitignore
@@ -172,3 +172,4 @@ cython_debug/
 .vscode/
 
 .gitignore
+models/
