+++ tests/test_base_model.py
@@ -4,6 +4,8 @@ import asyncpg
 from tuskorm.models.base_model import BaseModel
 from unittest.mock import patch
 import subprocess
+import os
+import json
 
 
 @pytest.mark.asyncio
@@ -26,15 +28,28 @@ async def test_handle_db_exception(caplog):
 
 
 @pytest.mark.asyncio
-async def test_generate_models_no_db_connection(monkeypatch):
-    """Test handling of database connection failure."""
-    monkeypatch.setenv("TUSK_DB_PORT", "9999")
-
+async def test_generate_models_no_db_connection(monkeypatch, tmp_path):
+    # Create a temporary .DBConfig with a bad port
+# Create a temporary .DBConfig with a bad port
+    bad_config = {
+        "host": "localhost",
+        "port": 9999,  # Invalid port for connection
+        "username": "tuskorm",
+        "password": "tuskorm",
+        "database": "tuskorm_test"
+    }
+    config_file = tmp_path / ".DBConfig"
+    config_file.write_text(json.dumps(bad_config))
+    
+    # Build the full path to tusk.py relative to the tests directory
+    script_path = os.path.join(os.path.dirname(__file__), "..", "tusk.py")
+    
+    # Change working directory so that tusk.py can pick up the temporary .DBConfig file
+    monkeypatch.chdir(tmp_path)
+    
     result = subprocess.run(
-        ["python", "tusk.py", "generate_models"], capture_output=True, text=True
+        ["python", script_path, "generate_models"],
+        capture_output=True, text=True
     )
-
+    
     assert result.returncode == 1, "Expected exit code 1 for failed DB connection"
-    assert (
-        "Could not connect to database" in result.stderr
-    ), "Expected database connection error message not found."
